separate suction voxel matrix from grasp data
